#835 Move TimeControl to Basics
