+++ CLAUDE.md
@@ -1,0 +1,7 @@
+# MMGIS - Claude Code Context
+
+This file provides context to Claude Code about the MMGIS project.
+
+**For complete project context, see AGENTS.md.**
+
+@AGENTS.md
