+++ src/essence/Ancillary/Coordinates.js
@@ -3,7 +3,7 @@ import $ from 'jquery'
 import F_ from '../Basics/Formulae_/Formulae_'
 import Map_ from '../Basics/Map_/Map_'
 import L_ from '../Basics/Layers_/Layers_'
-import TimeUI from './TimeUI'
+import TimeUI from '../Basics/TimeControl_/TimeUI'
 import Dropy from '../../external/Dropy/dropy'
 import * as _UserInterface_ from '../Basics/UserInterface_/UserInterface_'
 import calls from '../../pre/calls'
