+++ src/essence/Tools/Animation/AnimationTool.js
@@ -7,7 +7,7 @@ import Viewer_ from '../../Basics/Viewer_/Viewer_'
 import ToolController_ from '../../Basics/ToolController_/ToolController_'
 import CursorInfo from '../../Ancillary/CursorInfo'
 import Description from '../../Ancillary/Description'
-import TimeControl from '../../Ancillary/TimeControl'
+import TimeControl from '../../Basics/TimeControl_/TimeControl'
 import Modal from '../../Ancillary/Modal'
 import HTML2Canvas from 'html2canvas'
 import gifshot from 'gifshot'
