+++ src/essence/essence.js
@@ -39,7 +39,7 @@ import MapLogo from './Ancillary/MapLogo'
 import Attributions from './Ancillary/Attributions'
 //import Swap from './Ancillary/Swap'
 import QueryURL from './Ancillary/QueryURL'
-import TimeControl from './Ancillary/TimeControl'
+import TimeControl from './Basics/TimeControl_/TimeControl'
 import calls from '../pre/calls'
 import { mmgisAPI_, mmgisAPI } from './mmgisAPI/mmgisAPI'
 import { makeMissionNotFoundDiv } from './LandingPage/LandingPage'
